+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsUpdateManagementPolicy/MSFT_TeamsUpdateManagementPolicy.psm1
@@ -230,34 +230,19 @@ function Set-TargetResource
         -InboundParameters $PSBoundParameters
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
+    $PSBoundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
 
     if ($CurrentValues.Ensure -eq 'Absent' -and $Ensure -eq 'Present')
     {
         Write-Verbose "Creating new Teams Update Management Policy {$Identity}"
-        $newParams = $PSBoundParameters
-        $newParams.Remove('Ensure') | Out-Null
-        $newParams.Remove('Credential') | Out-Null
-        $newParams.Remove('ApplicationId') | Out-Null
-        $newParams.Remove('TenantId') | Out-Null
-        $newParams.Remove('CertificateThumbprint') | Out-Null
-        $newParams.Remove('ManagedIdentity') | Out-Null
-        $newParams.Remove('AccessTokens') | Out-Null
 
-        New-CsTeamsUpdateManagementPolicy @newParams | Out-Null
+        New-CsTeamsUpdateManagementPolicy @PSBoundParameters | Out-Null
     }
     elseif ($CurrentValues.Ensure -eq 'Present' -and $Ensure -eq 'Present')
     {
         Write-Verbose "Updating existing Teams Update Management Policy {$Identity}"
-        $setParams = $PSBoundParameters
-        $setParams.Remove('Ensure') | Out-Null
-        $setParams.Remove('Credential') | Out-Null
-        $setParams.Remove('ApplicationId') | Out-Null
-        $setParams.Remove('TenantId') | Out-Null
-        $setParams.Remove('CertificateThumbprint') | Out-Null
-        $setParams.Remove('ManagedIdentity') | Out-Null
-        $setParams.Remove('AccessTokens') | Out-Null
 
-        Set-CsTeamsUpdateManagementPolicy @setParams | Out-Null
+        Set-CsTeamsUpdateManagementPolicy @PSBoundParameters | Out-Null
     }
     elseif ($CurrentValues.Ensure -eq 'Present' -and $Ensure -eq 'Absent')
     {
@@ -355,16 +340,27 @@ function Test-TargetResource
     Write-Verbose -Message "Testing configuration of Team Update Management Policy {$Identity}"
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
+    $ValuesToCheck = ([Hashtable]$PSBoundParameters).Clone()
+
+    if ($CurrentValues.Ensure -ne $Ensure)
+    {
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+    $testResult = $true
+
+    $ValuesToCheck = Remove-M365DSCAuthenticationParameter -BoundParameters $ValuesToCheck
 
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $PSBoundParameters)"
 
-    $ValuesToCheck = $PSBoundParameters
-
-    $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
-        -Source $($MyInvocation.MyCommand.Source) `
-        -DesiredValues $PSBoundParameters `
-        -ValuesToCheck $ValuesToCheck.Keys
+    if ($testResult)
+    {
+        $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
+            -Source $($MyInvocation.MyCommand.Source) `
+            -DesiredValues $PSBoundParameters `
+            -ValuesToCheck $ValuesToCheck.Keys
+    }
 
     Write-Verbose -Message "Test-TargetResource returned $TestResult"